requirements.txt
aiogram>=3.3,<4
fastapi>=0.111
uvicorn>=0.29
python-dotenv>=1.0
pydantic>=2.6
